+++ web/package.json
@@ -71,6 +71,7 @@
     "react-dom": "18.2.0",
     "react-expanding-textarea": "2.3.6",
     "react-hot-toast": "2.2.0",
+    "react-lottie": "1.2.3",
     "react-masonry-css": "1.0.16",
     "react-plotly.js": "2.6.0",
     "react-query": "3.39.0",
@@ -90,6 +91,7 @@
     "@types/lodash": "4.14.178",
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.9",
+    "@types/react-lottie": "1.2.6",
     "@types/string-similarity": "^4.0.0",
     "autoprefixer": "10.2.6",
     "cross-env": "^7.0.3",
